app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
